requirements.txt
# Main Application Requirements
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
redis>=4.0.0
pydantic>=1.8.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=0.19.0
librosa>=0.9.0
soundfile>=0.10.3
numpy>=1.21.0
tensorflow>=2.8.0

# Training Module Requirements
# Note: These are only needed for training the model
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
sox>=1.4.0  # For audio conversion
pydub>=0.25.1  # For audio processing

# Testing Dependencies
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
requests==2.31.0 